backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic>=2.0.0,<3.0.0
python-dotenv==1.2.1
bcrypt==4.1.3
python-jose[cryptography]==3.5.0
python-multipart==0.0.21
email-validator==2.3.0
starlette==0.37.2
pillow>=10.0.0,<13.0.0
pywebpush==2.0.1
py-vapid==1.9.2
